pass back isanmating to loopview to show looping text as long as animation spins. Animate changing text lengths of content.
